mason was here

+++ AdminInfo/AdminPage.aspx.cs
@@ -15,6 +15,7 @@ namespace CSCI213.AdminInfo
     {
         string conn = "Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=C:\\Users\\mason.myles\\Desktop\\213Assignment\\CSCI213\\App_Data\\KarateSchool.mdf;Integrated Security=True;Connect Timeout=30";
         LogonDataContext dbcon;
+
         protected void Page_Load(object sender, EventArgs e)
         {
             dbcon = new LogonDataContext(conn);
@@ -28,8 +29,21 @@ namespace CSCI213.AdminInfo
 
                 DropDownList5.DataSource = result;
                 DropDownList5.DataBind();
+
+                var allMembers = from x in dbcon.Members
+                                 select new { MemName = x.MemberFirstName + "  " + x.MemberLastName, x.Member_UserID };
+                DropDownList6.DataTextField = "MemName";
+                DropDownList6.DataValueField = "Member_UserID";
+                DropDownList6.DataSource = allMembers;
+                DropDownList6.DataBind();
+
+                var allInstructors = from x in dbcon.Instructors
+                                 select new { InsName = x.InstructorFirstName + "  " + x.InstructorLastName, x.InstructorID };
+                DropDownList7.DataTextField = "InsName";
+                DropDownList7.DataValueField = "InstructorID";
+                DropDownList7.DataSource = allInstructors;
+                DropDownList7.DataBind();
             }
-            dbcon = new LogonDataContext(conn);
 
             //show table of all members
             var member = (from x in dbcon.Members
@@ -62,6 +76,14 @@ namespace CSCI213.AdminInfo
                 var memeberId = (from x in dbcon.NetUsers
                                  where x.UserName == userName.Text
                                  select x.UserID).Single();
+                                int instructorId = getInstructorId(DropDownList5, dbcon); 
+                decimal sectionFee = Convert.ToInt32(TextBox14.Text);
+                Section section = new Section();
+                section.SectionName = DropDownList3.SelectedValue;
+                section.SectionFee = sectionFee;
+                section.Member_ID = memeberId;
+                section.SectionStartDate = DateTime.Parse(TextBox6.Text);
+                section.Instructor_ID = instructorId;
 
                 Member mem = new Member();
                 TextBox1.Text = memeberId.ToString();
@@ -69,30 +91,15 @@ namespace CSCI213.AdminInfo
                 mem.MemberFirstName = TextBox2.Text;
                 mem.MemberLastName = TextBox3.Text;
                 mem.MemberPhoneNumber = TextBox4.Text;
-                mem.MemberDateJoined = DateTime.UtcNow;
+                mem.MemberDateJoined = DateTime.Parse(TextBox6.Text);
+
                 TextBox6.Text = DateTime.UtcNow.ToString();
                 mem.MemberEmail = TextBox5.Text;
 
-                int instructorId = getInstructorId(DropDownList5, dbcon); ;
-                Section section = new Section();
-                section.SectionName = DropDownList3.SelectedValue;
-                section.SectionFee = 0;
-                section.Member_ID = memeberId;
-                section.Instructor_ID = instructorId;
-
                 dbcon.Members.InsertOnSubmit(mem);
+                dbcon.Sections.InsertOnSubmit(section);
                 dbcon.SubmitChanges();
             }
-        }
-
-        private int getInstructorId(DropDownList dtr, LogonDataContext db)
-        {
-            int id = Convert.ToInt32(dtr.SelectedValue); ;
-
-            var memeberId = (from x in dbcon.NetUsers
-                             where x.UserID == id
-                             select x.UserID).Single();
-            return memeberId;
         }
 
         protected void Button3_Click(object sender, EventArgs e)
@@ -126,37 +133,61 @@ namespace CSCI213.AdminInfo
             dbcon.Instructors.InsertOnSubmit(instructor);
             dbcon.SubmitChanges();
             UpdateQuery();
+            updateAllDropDowns();
         }
 
         protected void Button4_Click(object sender, EventArgs e)
         {
-            string memberToDelete = TextBox12.Text;
-
-            var deletedNetUserID = (from y in dbcon.NetUsers
-                                    where y.UserName == memberToDelete
-                                    where y.UserType == "Member"
-                                    select y.UserID).Single();
+            int id = Convert.ToInt32(DropDownList6.SelectedValue);
 
             var deletedNetUser = (from y in dbcon.NetUsers
-                                  where y.UserID == deletedNetUserID
-                                  select y).Single();
+                                  where y.UserID == id
+                                  select y).SingleOrDefault();
 
             var deletedResult = (from x in dbcon.Members
-                                 where x.Member_UserID == deletedNetUserID
-                                 select x).Single();
+                                 where x.Member_UserID == id
+                                 select x).SingleOrDefault();
 
-            int sectionID = Convert.ToInt32(TextBox13.Text);
+            if(deletedNetUser != null)
+            {
+                dbcon.NetUsers.DeleteOnSubmit(deletedNetUser);
+            }
 
-            var alterSection = (from p in dbcon.Sections
-                                where p.Member_ID == deletedNetUserID
-                                where p.SectionID == sectionID
-                                select p).SingleOrDefault();
+            if(deletedResult != null)
+            {
+                dbcon.Members.DeleteOnSubmit(deletedResult);
+            }
 
-            dbcon.Members.DeleteOnSubmit(deletedResult);
-            dbcon.NetUsers.DeleteOnSubmit(deletedNetUser);
-            dbcon.Sections.DeleteOnSubmit(alterSection);
             dbcon.SubmitChanges();
             UpdateQuery();
+            updateAllDropDowns();
+        }
+
+        protected void Button5_Click(object sender, EventArgs e)
+        {
+            int id = Convert.ToInt32(DropDownList7.SelectedValue);
+
+            var deletedNetUser = (from y in dbcon.NetUsers
+                                  where y.UserID == id
+                                  select y).SingleOrDefault();
+
+            var deletedResult = (from x in dbcon.Instructors
+                                 where x.InstructorID == id
+                                 select x).SingleOrDefault();
+
+            if(deletedNetUser != null)
+            {
+                dbcon.Instructors.DeleteOnSubmit(deletedResult);
+            }
+
+            if(deletedResult != null)
+            {
+                dbcon.NetUsers.DeleteOnSubmit(deletedNetUser);
+            }
+
+            dbcon.SubmitChanges();
+            UpdateQuery();
+            updateAllDropDowns();
         }
 
         protected void Button6_Click(object sender, EventArgs e)
@@ -181,38 +212,44 @@ namespace CSCI213.AdminInfo
             InstructorsGridView.DataBind();
         }
 
-        protected void Button5_Click(object sender, EventArgs e)
+        private void updateAllDropDowns()
         {
-            string memberToDelete = TextBox11.Text;
+            if(IsPostBack)
+            {
+                dbcon = new LogonDataContext(conn);
+                var result = from item in dbcon.Instructors
+                             orderby item.InstructorFirstName
+                             select new { Instructor = item.InstructorFirstName, item.InstructorID };
+                DropDownList5.DataTextField = "Instructor";
+                DropDownList5.DataValueField = "InstructorId";
 
-            var deletedNetUserID = (from y in dbcon.NetUsers
-                                    where y.UserName == memberToDelete
-                                    where y.UserType == "Instructor"
-                                    select y.UserID).Single();
+                DropDownList5.DataSource = result;
+                DropDownList5.DataBind();
 
-            var deletedNetUser = (from y in dbcon.NetUsers
-                                  where y.UserName == memberToDelete
-                                  where y.UserType == "Instructor"
-                                  select y).Single();
+                var allMembers = from x in dbcon.Members
+                                 select new { MemName = x.MemberFirstName + "  " + x.MemberLastName, x.Member_UserID };
+                DropDownList6.DataTextField = "MemName";
+                DropDownList6.DataValueField = "Member_UserID";
+                DropDownList6.DataSource = allMembers;
+                DropDownList6.DataBind();
 
-            var deletedResult = (from x in dbcon.Instructors
-                                 where x.InstructorID == deletedNetUserID
-                                 select x).Single();
+                var allInstructors = from x in dbcon.Instructors
+                                     select new { InsName = x.InstructorFirstName + "  " + x.InstructorLastName, x.InstructorID };
+                DropDownList7.DataTextField = "InsName";
+                DropDownList7.DataValueField = "InstructorID";
+                DropDownList7.DataSource = allInstructors;
+                DropDownList7.DataBind();
+            }
+        }
 
+        private int getInstructorId(DropDownList dtr, LogonDataContext db)
+        {
+            int id = Convert.ToInt32(dtr.SelectedValue); ;
 
-
-            int sectionID = Convert.ToInt32(TextBox13.Text);
-
-            var alterSection = (from p in dbcon.Sections
-                                where p.Member_ID == deletedNetUserID
-                                where p.SectionID == sectionID
-                                select p).SingleOrDefault();
-
-            dbcon.Instructors.DeleteOnSubmit(deletedResult);
-            dbcon.NetUsers.DeleteOnSubmit(deletedNetUser);
-            dbcon.Sections.DeleteOnSubmit(alterSection);
-            dbcon.SubmitChanges();
-            UpdateQuery();
+            var memeberId = (from x in db.NetUsers
+                             where x.UserID == id
+                             select x.UserID).Single();
+            return memeberId;
         }
     }
 }
